+++ tests/E2E/AssetLocaleTest.php
@@ -6,6 +6,7 @@
 
 namespace Contentful\Tests\E2E;
 
+use Contentful\File\File;
 use Contentful\Delivery\Client;
 use Contentful\Delivery\Query;
 
@@ -40,5 +41,13 @@ class AssetLocaleTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals('es', $asset->getLocale());
         $this->assertEquals('ben-chang', $asset->getTitle());
         $this->assertEquals('I AM A SPANISH GENIUS!', $asset->getDescription());
+
+        // Tests that the fallback chain is working correctly.
+        // Tested chain is es -> it -> en-US
+        $this->assertNull($asset->getFile('it'));
+        $this->assertNull($asset->getFile('en-US'));
+        $this->assertInstanceOf(File::class, $asset->getFile());
+        $this->assertInstanceOf(File::class, $asset->getFile('es'));
+        $this->assertEquals('//images.contentful.com/88dyiqcr7go8/SQOIQ1rZMQQUeyoyGiEUq/84b6aef287ed214b464114655f99bfa8/ben-chang.jpg', $asset->getFile('es')->getUrl());
     }
 }